Add files via upload
My version of the pipeline includes changing the minimap2 paramters to have a -f1000 flag to set the minimum fragment length to be 1000bp as we are working with long read data from Nanopore sequencing platforms. I have also modified the snakefile rules to include "script" arguments instead of "shell" arguments when using the python scripts as errors were being produced with the latest version of apogee.